requirements.txt
streamlit
embedchain
streamlit-chat
sentence-transformers
python-dotenv
openai